% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{A new project function}
\usage{
create_project(name = "my_project", type = "pdf")
}
\arguments{
\item{name}{A name for your project}

\item{type}{A valid .Rmd output (html, pdf, word)}
}
\description{
This function will create a new project directory for you
with subdirectories, and .Rmd and .R templates.
}
\examples{
# Create a new project with html report
# create_project(name = 'my_project', type = 'pdf')
}
\keyword{projects}
