% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_n_digits.R
\name{give_n_digits}
\alias{give_n_digits}
\title{Give N digits}
\usage{
give_n_digits(x, n = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{the number of decimal places to print}
}
\description{
A helper function that always prints a specified number of decimals
}
\examples{
give_n_digits(0.0293526, n = 2)
}
\keyword{helpers}
\keyword{printing}
