% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_to_table.R
\name{mod_to_table}
\alias{mod_to_table}
\title{Model to table}
\usage{
mod_to_table(model, param_names = NULL, font_size = 11,
  left_align = 1, width = NULL, col = NULL)
}
\arguments{
\item{model}{A model object}

\item{param_names}{For renaming the default parameter names}

\item{font_size}{The font size for the table}

\item{left_align}{Columns that should be left aligned}

\item{width}{The width in inches of the columns}

\item{col}{The number of a specific column for width adjustments}
}
\description{
This function takes a model object generates a table
}
\examples{
mod <- lm(mpg ~ drat, data = mtcars)
mod_to_table(mod)
}
\keyword{Report}
\keyword{model}
