% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_approximation.R
\name{normal_approximation}
\alias{normal_approximation}
\title{Calculate p-values using the normal approximation}
\usage{
normal_approximation(statistic)
}
\arguments{
\item{statistic}{The estimated statistic for a model parameter}
}
\description{
This function calculates a p-value from a t-value
}
\examples{
library(lme4)
mod3a <- lmer(Reaction ~ 1 + Days + (1|Subject), data = sleepstudy)
normal_approximation(2.02)
}
\keyword{Report}
\keyword{model}
