% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_builder.R
\name{print_builder}
\alias{print_builder}
\title{Print builder}
\usage{
print_builder(model, latex = TRUE)
}
\arguments{
\item{model}{a model object}

\item{latex}{Output is LaTeX or markdown}
}
\description{
A series of helper functions for printing model parameters
}
\keyword{Report}
\keyword{model}
