% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dir_ul.R
\name{print_dir_ul}
\alias{print_dir_ul}
\title{An function to generate unordered lists from the directory structure}
\usage{
print_dir_ul(path = ".", remove = NULL, nested = FALSE)
}
\arguments{
\item{path}{The path to the directory of which you want to print the
structure}

\item{remove}{A regular expression to subset the output}

\item{nested}{(logical) If false, creates unordered list of current
directory only. Otherwise, subdirectories within current directory
(1 level) are also included in the list.}
}
\description{
This function will create an unordered list of all the files in a directory
and print the output. It should be used in conjunction with RMarkdown and
knitr. Specifically, you want to call it within an R chunk that has the
results set to 'asis'.
}
\examples{
# In any directory
print_dir_ul()

# To specify the path
print_dir_ul(path = "./")

# All files that are not .txt
print_dir_ul(remove = ".txt")
}
\keyword{list}
\keyword{unordered}
