% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_param.R
\name{print_model_param}
\alias{print_model_param}
\title{Print model parameters}
\usage{
print_model_param(model, parameter, latex = TRUE)
}
\arguments{
\item{model}{A model object}

\item{parameter}{A model parameter}

\item{latex}{Determine if output is LaTeX or markdown}
}
\description{
This function takes a model object and prints a row. This is used for
reporting results.
}
\keyword{Report}
\keyword{model}
