% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_params.R
\name{print_model_params}
\alias{print_model_params}
\title{Helper function for reporting individual model parameters}
\usage{
print_model_params(model, predictor, parameter)
}
\arguments{
\item{model}{A model object}

\item{predictor}{(character) The effect (fixed/random) you want to report}

\item{parameter}{(character) The parameter you want to report ('estimate',
'std.error', 'statistic')}
}
\description{
This function takes a model object, a predictor, and a parameter and prints
the numeric output.
}
\examples{
library(lme4)
mod1 <- lm(mpg ~ wt, data = mtcars)
mod2 <- lmer(Reaction ~ 1 + Days + (1|Subject), data = sleepstudy,
             REML = FALSE)
print_model_params(model = mod1, predictor = "wt", parameter = "estimate")
print_model_params(model = mod2, predictor = "Days", parameter = "estimate")
}
\keyword{Report}
\keyword{parameter}
