% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_params_p.R
\name{print_model_params_p}
\alias{print_model_params_p}
\title{Helper function for reporting model parameters in parenthesis}
\usage{
print_model_params_p(model, predictor, latex = TRUE)
}
\arguments{
\item{model}{A model object}

\item{predictor}{(character) The effect (fixed/random) you want to report}

\item{latex}{(Logical) If true, output is appropriate for LaTeX. Otherwise
formatting is for markdown.}
}
\description{
This function takes a model object  and a predictor, and prints the output
in parenthesis. Includes confidence intervals, stat type and p-values.
}
\examples{
library(lme4)
library(lmerTest)
mod1 <- lm(mpg ~ wt, data = mtcars)
mod2 <- lmer(Reaction ~ 1 + Days + (1|Subject), data = sleepstudy,
             REML = FALSE)
print_model_params_p(model = mod1, predictor = "wt", latex = FALSE)
print_model_params_p(model = mod2, predictor = "Days")
}
\keyword{Report}
\keyword{model}
\keyword{parameter}
