% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nmc_p.R
\name{print_nmc_p}
\alias{print_nmc_p}
\title{Helper function for reporting nested model comparisons}
\usage{
print_nmc_p(table, model, latex = TRUE)
}
\arguments{
\item{table}{An ANOVA table}

\item{model}{The model you want to report}

\item{latex}{(Logical) If true, output is appropriate for LaTeX. Otherwise
formatting is for markdown.}
}
\description{
This function takes a nested model comparison ANOVA object and prints the
result in LaTeX or markdown. This currenlty only works for linear mixed
effects model comparisons.
}
\examples{
library(lme4)
mod1 <- lmer(Reaction ~ 1 + (1|Subject), data = sleepstudy, REML = FALSE)
mod2 <- lmer(Reaction ~ 1 + Days + (1|Subject), data = sleepstudy,
             REML = FALSE)
aov_table <- anova(mod1, mod2, test = "Chisq")
print_nmc_p(table = aov_table, model = 'mod2')
print_nmc_p(table = aov_table, model = 'mod2', latex = FALSE)
}
\keyword{Report}
\keyword{comparisons}
\keyword{model}
\keyword{nested}
