% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pval.R
\name{print_pval}
\alias{print_pval}
\title{A helper function to round p-values and paste less than or greater than
symbol}
\usage{
print_pval(x, latex = TRUE)
}
\arguments{
\item{x}{(numeric) A p-value to round}

\item{latex}{(Logical) If true, output is appropriate for LaTeX. Otherwise
formatting is for markdown.}
}
\description{
This function takes a numeric value and rounds it, then it formats it in
order to be reported in an academic report written in LaTeX or markdown.
It is most useful for R output that is more than three decimal places long.
}
\examples{
print_pval(0.03765)
print_pval(0.0000001, latex = FALSE)
}
\keyword{Round}
\keyword{p-values}
