% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printP.R
\name{printP}
\alias{printP}
\title{A helper function to round p-values and paste less than or greater than
symbol}
\usage{
printP(x, latex = TRUE)
}
\arguments{
\item{x}{A p-value to round}

\item{latex}{(Logical) If true, output is appropriate for LaTeX. Otherwise
formatting is for markdown.}
}
\description{
This function takes a numeric value and rounds it, then it formats it in
order to be reported in an academic report written in LaTeX or markdown.
It is most useful for R output that is more than three decimal places long.
}
\examples{
printP(0.03765)
printP(0.0000001, latex = FALSE)
}
\keyword{Round}
\keyword{p-values}
