% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_pval.R
\name{round_pval}
\alias{round_pval}
\title{A rounding function for pvalues}
\usage{
round_pval(x)
}
\arguments{
\item{x}{(numeric) A p-value to round}
}
\description{
This function takes a numeric value and rounds it. It is most useful for
R output that is more than three decimal places long.
}
\examples{
round_pval(0.03765)
round_pval(0.0000001)
}
\keyword{Round}
\keyword{p-values}
