% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundP.R
\name{roundP}
\alias{roundP}
\title{A rounding function for pvalues}
\usage{
roundP(x)
}
\arguments{
\item{x}{(numeric) A p-value to round}
}
\description{
This function takes a numeric value and rounds it. It is most useful for
R output that is more than three decimal places long.
}
\examples{
roundP(0.03765)
roundP(0.0000001)
}
\keyword{Round}
\keyword{p-values}
